@@ -10,13 +10,15 @@
     "install:clean": "rm -rf node_modules/ && rm -rf yarn.lock && rm -rf .next/ && yarn install && yarn dev"
   },
   "dependencies": {
+    "@emotion/cache": "latest",
+    "@emotion/css": "^11.10.0",
     "@emotion/react": "^11.10.4",
+    "@emotion/server": "^11.10.0",
     "@emotion/styled": "^11.10.4",
-    "@material-ui/core": "4.11.3",
-    "@material-ui/icons": "4.11.2",
     "@mui/icons-material": "^5.10.6",
     "@mui/lab": "^5.0.0-alpha.101",
     "@mui/material": "^5.10.6",
+    "@mui/styles": "^5.10.8",
     "axios": "^0.27.2",
     "chartist": "0.10.1",
     "classnames": "2.2.6",
@@ -33,11 +35,11 @@
     "next-redux-wrapper": "^8.0.0",
     "nookies": "^2.5.2",
     "path": "0.12.7",
-    "perfect-scrollbar": "1.5.0",
     "prop-types": "15.7.2",
-    "react": "17.0.1",
+    "react": "^18.2.0",
     "react-chartist": "0.14.3",
-    "react-dom": "17.0.1",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
     "react-pdf": "^5.7.2",
     "react-redux": "^8.0.2",
     "react-swipeable-views": "0.13.9",
