@@ -1,7 +1,7 @@
 /*eslint-disable*/
 import React from "react";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
+// @mui/core components
+import { makeStyles } from "@mui/core/styles";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/footerStyle.js";
 
