@@ -1,9 +1,9 @@
 import React from "react";
 // nodejs library to set properties for components
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import Grid from "@material-ui/core/Grid";
+// @mui/core components
+import { makeStyles } from "@mui/core/styles";
+import Grid from "@mui/core/Grid";
 
 const styles = {
   grid: {
