@@ -1,12 +1,12 @@
 import React from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import Snack from "@material-ui/core/SnackbarContent";
-import IconButton from "@material-ui/core/IconButton";
-// @material-ui/icons
-import Close from "@material-ui/icons/Close";
+// @mui/core components
+import { makeStyles } from "@mui/core/styles";
+import Snack from "@mui/core/SnackbarContent";
+import IconButton from "@mui/core/IconButton";
+// @mui/icons
+import Close from "@mui/icons/Close";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/snackbarContentStyle.js";
 
