@@ -1,12 +1,12 @@
 import React from "react";
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import Table from "@material-ui/core/Table";
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
+// @mui/core components
+import { makeStyles } from "@mui/core/styles";
+import Table from "@mui/core/Table";
+import TableHead from "@mui/core/TableHead";
+import TableRow from "@mui/core/TableRow";
+import TableBody from "@mui/core/TableBody";
+import TableCell from "@mui/core/TableCell";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/tableStyle.js";
 import { useEffect } from "react";
