@@ -1,19 +1,19 @@
 import React from "react";
 import PropTypes from "prop-types";
 import classnames from "classnames";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import Checkbox from "@material-ui/core/Checkbox";
-import Tooltip from "@material-ui/core/Tooltip";
-import IconButton from "@material-ui/core/IconButton";
-import Table from "@material-ui/core/Table";
-import TableRow from "@material-ui/core/TableRow";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-// @material-ui/icons
-import Edit from "@material-ui/icons/Edit";
-import Close from "@material-ui/icons/Close";
-import Check from "@material-ui/icons/Check";
+// @mui/core components
+import { makeStyles } from "@mui/core/styles";
+import Checkbox from "@mui/core/Checkbox";
+import Tooltip from "@mui/core/Tooltip";
+import IconButton from "@mui/core/IconButton";
+import Table from "@mui/core/Table";
+import TableRow from "@mui/core/TableRow";
+import TableBody from "@mui/core/TableBody";
+import TableCell from "@mui/core/TableCell";
+// @mui/icons
+import Edit from "@mui/icons/Edit";
+import Close from "@mui/icons/Close";
+import Check from "@mui/icons/Check";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/tasksStyle.js";
 
