@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
+// @mui/core components
+import { makeStyles } from "@mui/core/styles";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/typographyStyle.js";
 
