@@ -1,21 +1,21 @@
 import React from "react";
 // react plugin for creating charts
 import ChartistGraph from "react-chartist";
-// @material-ui/core
-import { makeStyles } from "@material-ui/core/styles";
-import Icon from "@material-ui/core/Icon";
-// @material-ui/icons
-import Store from "@material-ui/icons/Store";
-import Warning from "@material-ui/icons/Warning";
-import DateRange from "@material-ui/icons/DateRange";
-import LocalOffer from "@material-ui/icons/LocalOffer";
-import Update from "@material-ui/icons/Update";
-import ArrowUpward from "@material-ui/icons/ArrowUpward";
-import AccessTime from "@material-ui/icons/AccessTime";
-import Accessibility from "@material-ui/icons/Accessibility";
-import BugReport from "@material-ui/icons/BugReport";
-import Code from "@material-ui/icons/Code";
-import Cloud from "@material-ui/icons/Cloud";
+// @mui/core
+import { makeStyles } from "@mui/core/styles";
+import Icon from "@mui/core/Icon";
+// @mui/icons
+import Store from "@mui/icons/Store";
+import Warning from "@mui/icons/Warning";
+import DateRange from "@mui/icons/DateRange";
+import LocalOffer from "@mui/icons/LocalOffer";
+import Update from "@mui/icons/Update";
+import ArrowUpward from "@mui/icons/ArrowUpward";
+import AccessTime from "@mui/icons/AccessTime";
+import Accessibility from "@mui/icons/Accessibility";
+import BugReport from "@mui/icons/BugReport";
+import Code from "@mui/icons/Code";
+import Cloud from "@mui/icons/Cloud";
 // layout for this page
 import Admin from "layouts/Admin.js";
 // core components
@@ -106,8 +106,8 @@ function Dashboard() {
         </GridItem>
         <GridItem xs={12} sm={6} md={3}>
           <Card>
-            <CardHeader color="info" stats icon>
-              <CardIcon color="info">
+            <CardHeader  stats icon>
+              <CardIcon >
                 <Accessibility />
               </CardIcon>
               <p className={classes.cardCategory}>Followers</p>
