@@ -1,155 +1,63 @@
-import React from "react";
-// react plugin for creating charts
-import ChartistGraph from "react-chartist";
-// @material-ui/core
-import { makeStyles } from "@material-ui/core/styles";
-import Icon from "@material-ui/core/Icon";
-// @material-ui/icons
-import Store from "@material-ui/icons/Store";
-import Warning from "@material-ui/icons/Warning";
-import DateRange from "@material-ui/icons/DateRange";
-import LocalOffer from "@material-ui/icons/LocalOffer";
-import Update from "@material-ui/icons/Update";
-import ArrowUpward from "@material-ui/icons/ArrowUpward";
-import AccessTime from "@material-ui/icons/AccessTime";
-import Accessibility from "@material-ui/icons/Accessibility";
-import History from "@material-ui/icons/History";
-// core components
-import GridItem from "components/Grid/GridItem.js";
-import GridContainer from "components/Grid/GridContainer.js";
-import Table from "components/Table/Table.js";
-import Tasks from "components/Tasks/Tasks.js";
-import CustomTabs from "components/CustomTabs/CustomTabs.js";
-import Danger from "components/Typography/Danger.js";
-import Card from "components/Card/Card.js";
-import CardHeader from "components/Card/CardHeader.js";
-import CardIcon from "components/Card/CardIcon.js";
-import CardBody from "components/Card/CardBody.js";
-import CardFooter from "components/Card/CardFooter.js";
-
-import { bugs, website } from "variables/general.js";
-
-import {
-  dailySalesChart,
-  emailsSubscriptionChart,
-  completedTasksChart,
-} from "variables/charts.js";
-
-import styles from "assets/jss/nextjs-material-dashboard/views/dashboardStyle.js";
-import { AdfScannerRounded } from "@mui/icons-material";
+import React, { useEffect, useMemo, useState } from "react";
+import { Grid, Typography } from "@mui/material";
+import { useAccount } from "lib/contexts/accountContext";
+import { generateScannerDataTable } from "lib/scannerDataTable";
+import useSWR, { mutate } from "swr";
+import { fetchData } from "lib/fetch";
+import ScannerListContainer from "./ScannerListContainer";
 
 function Dashboard() {
-  const useStyles = makeStyles(styles);
-  const classes = useStyles();
+  const { scannerList, setScannerList } = useAccount();
+
+  const rowsPerPage = 6;
+  const [rowCount, setRowCount] = useState(0);
+  const [pageIndex, setPageIndex] = useState(1);
+  const scannerTableData = useMemo(() => 
+    generateScannerDataTable(scannerList ?? [], () => {
+      mutate(`${process.env.backendUrl}api/scanners?page=${pageIndex}&limit=${rowsPerPage}&sort=-lastActive`)
+    }), 
+    [scannerList]
+  );
+
+  const handlePageIndexChange = (e, newIndex) => setPageIndex(newIndex+1);
+
+  const { data, error, isValidating } = useSWR(
+    `${process.env.backendUrl}api/scanners?page=${pageIndex}&limit=${rowsPerPage}&sort=-lastActive`,
+    fetchData,
+    {
+      refreshInterval: 5000
+    }
+  );
+
+  useEffect(() => {
+    if (data) {
+      setScannerList(data?.data ?? []);
+      setRowCount(data?.dataCount ?? 0);
+    }
+  }, [data]);
+
   return (
-    <div>
-      <GridContainer>
-        <GridItem xs={12} sm={12} md={4}>
-          <Card chart>
-            <CardHeader color="success">
-              <ChartistGraph
-                className="ct-chart"
-                data={dailySalesChart.data}
-                type="Line"
-                options={dailySalesChart.options}
-                listener={dailySalesChart.animation}
-              />
-            </CardHeader>
-            <CardBody>
-              <h4 className={classes.cardTitle}>Daily Sales</h4>
-              <p className={classes.cardCategory}>
-                <span className={classes.successText}>
-                  <ArrowUpward className={classes.upArrowCardCategory} /> 55%
-                </span>{" "}
-                increase in today sales.
-              </p>
-            </CardBody>
-            <CardFooter chart>
-              <div className={classes.stats}>
-                <AccessTime /> updated 4 minutes ago
-              </div>
-            </CardFooter>
-          </Card>
-        </GridItem>
-        <GridItem xs={12} sm={12} md={4}>
-          <Card chart>
-            <CardHeader color="warning">
-              <ChartistGraph
-                className="ct-chart"
-                data={emailsSubscriptionChart.data}
-                type="Bar"
-                options={emailsSubscriptionChart.options}
-                responsiveOptions={emailsSubscriptionChart.responsiveOptions}
-                listener={emailsSubscriptionChart.animation}
-              />
-            </CardHeader>
-            <CardBody>
-              <h4 className={classes.cardTitle}>Email Subscriptions</h4>
-              <p className={classes.cardCategory}>Last Campaign Performance</p>
-            </CardBody>
-            <CardFooter chart>
-              <div className={classes.stats}>
-                <AccessTime /> campaign sent 2 days ago
-              </div>
-            </CardFooter>
-          </Card>
-        </GridItem>
-        <GridItem xs={12} sm={12} md={4}>
-          <Card chart>
-            <CardHeader color="dark">
-              <ChartistGraph
-                className="ct-chart"
-                data={completedTasksChart.data}
-                type="Line"
-                options={completedTasksChart.options}
-                listener={completedTasksChart.animation}
-              />
-            </CardHeader>
-            <CardBody>
-              <h4 className={classes.cardTitle}>Completed Tasks</h4>
-              <p className={classes.cardCategory}>Last Campaign Performance</p>
-            </CardBody>
-            <CardFooter chart>
-              <div className={classes.stats}>
-                <AccessTime /> campaign sent 2 days ago
-              </div>
-            </CardFooter>
-          </Card>
-        </GridItem>
-      </GridContainer>
-      <GridContainer>
-        <GridItem xs={12}>
-          <CustomTabs
-            title="Scanner"
-            headerColor="dark"
-            tabs={[
-              {
-                tabName: "List",
-                tabIcon: AdfScannerRounded,
-                tabContent: (
-                  <Tasks
-                    checkedIndexes={[0, 3]}
-                    tasksIndexes={[0, 1, 2, 3]}
-                    tasks={bugs}
-                  />
-                ),
-              },
-              {
-                tabName: "History",
-                tabIcon: History,
-                tabContent: (
-                  <Tasks
-                    checkedIndexes={[0]}
-                    tasksIndexes={[0, 1]}
-                    tasks={website}
-                  />
-                ),
-              }
-            ]}
-          />
-        </GridItem>
-      </GridContainer>
-    </div>
+    <>
+      <Grid container sx={{ padding: '30px 0' }} spacing={2}>
+        <Grid item xs={12} display="flex" alignItems="center">
+            <Typography sx={{ fontWeight: 500, fontSize: '20px', lineHeight: '24px', color: '#190D29' }}>
+              Scanner List
+            </Typography>
+        </Grid>
+          
+        <Grid item xs={12} display="flex" alignItems="center">
+          <Typography sx={{ fontWeight: 500, fontSize: '16px', lineHeight: '19px', color: '#190D29' }}>
+            List of connected scanners
+          </Typography>
+        </Grid>
+
+        {scannerList.map((scanner) => (
+          <Grid item xs={12} md={6}>
+            <ScannerListContainer {...scanner} pageIndex={pageIndex} rowsPerPage={rowsPerPage} />
+          </Grid>
+        ))}
+      </Grid>
+    </>
   );
 }
 
